+++ alembic.ini
@@ -60,7 +60,7 @@ version_path_separator = os  # Use os.pathsep. Default configuration used for ne
 # are written from script.py.mako
 # output_encoding = utf-8
 
-sqlalchemy.url=postgresql://postgres:12345678@localhost:5432/postgres
+sqlalchemy.url=postgresql+asyncpg://postgres:123456@localhost:5432/postgres
 
 
 [post_write_hooks]
