--- tests/deleted.py
@@ -1,1 +1,0 @@
-
