--- tests/deleteds.py
@@ -1,1 +1,0 @@
-
